Moved logic for rendering TicketList from App.jsx into separate components/tickets/TicketList.jsx component | moved CSS for TicketList from App.css to components/tickets/Tickets.css | moved or updated imports | imported and rendered TicketList component inside the App component function
